+++ examples/distribus_multiple_hosts.js
@@ -1,15 +1,15 @@
 var distribus = require('distribus');
-var actors = require('../index');
+var eve = require('../index');
 
 var host1 = new distribus.Host();
 var host2 = new distribus.Host();
 
 // create two transports: one connected to host1, the other to host2
-var transport1 = new actors.DistribusTransport({host: host1});
-var transport2 = new actors.DistribusTransport({host: host2});
+var transport1 = new eve.DistribusTransport({host: host1});
+var transport2 = new eve.DistribusTransport({host: host2});
 
-var actor1;
-var actor2;
+var agent1;
+var agent2;
 
 host1.listen('localhost', 3000)
 
@@ -22,31 +22,31 @@ host1.listen('localhost', 3000)
     })
 
     .then(function () {
-      // actor 1 listens for messages containing 'hi' or 'hello' (case insensitive)
-      actor1 = new actors.Actor('actor1');
-      actor1.on(/hi|hello/i, function (from, message) {
+      // agent 1 listens for messages containing 'hi' or 'hello' (case insensitive)
+      agent1 = new eve.Agent('agent1');
+      agent1.on(/hi|hello/i, function (from, message) {
         console.log(from + ' said: ' + message);
 
         // reply to the greeting
         this.send(from, 'Hi ' + from + ', nice to meet you!');
       });
 
-      // connect the actor to transport1
-      return actor1.connect(transport1);
+      // connect the agent to transport1
+      return agent1.connect(transport1);
     })
 
     .then(function () {
-      // actor 2 listens for any message
-      actor2 = new actors.Actor('actor2');
-      actor2.on(/./, function (from, message) {
+      // agent 2 listens for any message
+      agent2 = new eve.Agent('agent2');
+      agent2.on(/./, function (from, message) {
         console.log(from + ' said: ' + message);
       });
 
-      // connect the actor to transport2
-      return actor2.connect(transport2);
+      // connect the agent to transport2
+      return agent2.connect(transport2);
     })
 
     .then(function () {
-      // once both are connected, send a message to actor 1
-      actor2.send('actor1', 'Hello actor1!');
+      // once both are connected, send a message to agent 1
+      agent2.send('agent1', 'Hello agent1!');
     });