+++ examples/greeting.js
@@ -1,25 +1,25 @@
-var actors = require('../index');
+var eve = require('../index');
 
-var transport = new actors.LocalTransport();
-var actor1 = new actors.Actor('actor1');
-var actor2 = new actors.Actor('actor2');
+var transport = new eve.LocalTransport();
+var agent1 = new eve.Agent('agent1');
+var agent2 = new eve.Agent('agent2');
 
-// actor 1 listens for messages containing 'hi' or 'hello' (case insensitive)
-actor1.on(/hi|hello/i, function (from, message) {
+// agent 1 listens for messages containing 'hi' or 'hello' (case insensitive)
+agent1.on(/hi|hello/i, function (from, message) {
   console.log(from + ' said: ' + message);
 
   // reply to the greeting
   this.send(from, 'Hi ' + from + ', nice to meet you!');
 });
 
-// actor 2 listens for any message
-actor2.on(/./, function (from, message) {
+// agent 2 listens for any message
+agent2.on(/./, function (from, message) {
   console.log(from + ' said: ' + message);
 });
 
-// connect both actors to the transport
-actor1.connect(transport);
-actor2.connect(transport);
+// connect both agents to the transport
+agent1.connect(transport);
+agent2.connect(transport);
 
-// send a message to actor 1
-actor2.send('actor1', 'Hello actor1!');
+// send a message to agent 1
+agent2.send('agent1', 'Hello agent1!');